requirements.txt
slackclient
requests==2.22.0
